+++ derive_builder_core/src/bindings.rs
@@ -1,4 +1,4 @@
-use RawTokens;
+use syn::{self, Path};
 
 /// Bindings to be used by the generated code.
 #[derive(Debug, Clone, Copy, Default)]
@@ -9,128 +9,75 @@ pub struct Bindings {
 
 impl Bindings {
     /// String type.
-    pub fn string_ty(&self) -> RawTokens<&'static str> {
-        RawTokens(if self.no_std {
-            ":: alloc :: string :: String"
+    pub fn string_ty(&self) -> Path {
+        syn::parse_str(if self.no_std {
+            "::alloc::string::String"
         } else {
-            ":: std :: string :: String"
-        })
+            "::std::string::String"
+        }).unwrap()
     }
 
     /// Result type.
-    pub fn result_ty(&self) -> RawTokens<&'static str> {
-        RawTokens(if self.no_std {
-            ":: core :: result :: Result"
+    pub fn result_ty(&self) -> Path {
+        syn::parse_str(if self.no_std {
+            "::core::result::Result"
         } else {
-            ":: std :: result :: Result"
-        })
+            "::std::result::Result"
+        }).unwrap()
     }
 
     /// Option type.
-    pub fn option_ty(&self) -> RawTokens<&'static str> {
-        RawTokens(if self.no_std {
-            ":: core :: option :: Option"
+    pub fn option_ty(&self) -> Path {
+        syn::parse_str(if self.no_std {
+            "::core::option::Option"
         } else {
-            ":: std :: option :: Option"
-        })
+            "::std::option::Option"
+        }).unwrap()
     }
 
     /// PhantomData type.
-    pub fn phantom_data_ty(&self) -> RawTokens<&'static str> {
-        RawTokens(if self.no_std {
-            ":: core :: marker :: PhantomData"
+    pub fn phantom_data_ty(&self) -> Path {
+        syn::parse_str(if self.no_std {
+            "::core::marker::PhantomData"
         } else {
-            ":: std :: marker :: PhantomData"
-        })
+            "::std::marker::PhantomData"
+        }).unwrap()
     }
 
     /// Default trait.
-    pub fn default_trait(&self) -> RawTokens<&'static str> {
-        RawTokens(if self.no_std {
-            ":: core :: default :: Default"
+    pub fn default_trait(&self) -> Path {
+        syn::parse_str(if self.no_std {
+            "::core::default::Default"
         } else {
-            ":: std :: default :: Default"
-        })
+            "::std::default::Default"
+        }).unwrap()
     }
 
     /// Clone trait.
-    pub fn clone_trait(&self) -> RawTokens<&'static str> {
-        RawTokens(if self.no_std {
-            ":: core :: clone :: Clone"
+    pub fn clone_trait(&self) -> Path {
+        syn::parse_str(if self.no_std {
+            "::core::clone::Clone"
         } else {
-            ":: std :: clone :: Clone"
-        })
+            "::std::clone::Clone"
+        }).unwrap()
     }
 
     /// Into trait.
     #[cfg_attr(feature = "cargo-clippy", allow(wrong_self_convention))]
-    pub fn into_trait(&self) -> RawTokens<&'static str> {
-        RawTokens(if self.no_std {
-            ":: core :: convert :: Into"
+    pub fn into_trait(&self) -> Path {
+        syn::parse_str(if self.no_std {
+            "::core::convert::Into"
         } else {
-            ":: std :: convert :: Into"
-        })
+            "::std::convert::Into"
+        }).unwrap()
     }
 
     /// TryInto trait.
-    pub fn try_into_trait(&self) -> RawTokens<&'static str> {
-        RawTokens(if self.no_std {
-            ":: core :: convert :: TryInto"
+    pub fn try_into_trait(&self) -> Path {
+        syn::parse_str(if self.no_std {
+            "::core::convert::TryInto"
         } else {
-            ":: std :: convert :: TryInto"
-        })
+            "::std::convert::TryInto"
+        }).unwrap()
     }
-}
-
-#[test]
-fn std() {
-    let b = Bindings { no_std: false };
-
-    assert_eq!(b.string_ty().to_tokens(), quote!(::std::string::String));
-
-    assert_eq!(b.result_ty().to_tokens(), quote!(::std::result::Result));
-
-    assert_eq!(b.option_ty().to_tokens(), quote!(::std::option::Option));
-
-    assert_eq!(
-        b.phantom_data_ty().to_tokens(),
-        quote!(::std::marker::PhantomData)
-    );
-
-    assert_eq!(
-        b.default_trait().to_tokens(),
-        quote!(::std::default::Default)
-    );
-
-    assert_eq!(b.clone_trait().to_tokens(), quote!(::std::clone::Clone));
-
-    assert_eq!(b.into_trait().to_tokens(), quote!(::std::convert::Into));
-}
-
-#[test]
-fn no_std() {
-    let b = Bindings { no_std: true };
-
-    assert_eq!(
-        b.string_ty().to_tokens(),
-        quote!(::alloc::string::String)
-    );
-
-    assert_eq!(b.result_ty().to_tokens(), quote!(::core::result::Result));
-
-    assert_eq!(b.option_ty().to_tokens(), quote!(::core::option::Option));
-
-    assert_eq!(
-        b.phantom_data_ty().to_tokens(),
-        quote!(::core::marker::PhantomData)
-    );
-
-    assert_eq!(
-        b.default_trait().to_tokens(),
-        quote!(::core::default::Default)
-    );
-
-    assert_eq!(b.clone_trait().to_tokens(), quote!(::core::clone::Clone));
-
-    assert_eq!(b.into_trait().to_tokens(), quote!(::core::convert::Into));
 }