+++ derive_builder_core/src/doc_comment.rs
@@ -1,4 +1,5 @@
-use syn;
+use syn::Attribute;
+use syn::synom::Parser;
 
 /// Doc-comment, implementing `quote::ToTokens`.
 ///
@@ -16,17 +17,11 @@ use syn;
 /// #    let doc_comment = doc_comment_from("foo".to_string());
 /// #
 /// #    assert_eq!(quote!(#doc_comment), quote!(
-/// #[doc = r##"foo"##]
+/// #[doc = "foo"]
 /// #    ));
 /// # }
 /// ```
-pub fn doc_comment_from(s: String) -> syn::Attribute {
-    syn::Attribute {
-        style: syn::AttrStyle::Outer,
-        value: syn::MetaItem::NameValue(
-            syn::Ident::new("doc"),
-            syn::Lit::Str(s, syn::StrStyle::Raw(2)),
-        ),
-        is_sugared_doc: false,
-    }
+pub fn doc_comment_from(s: String) -> Attribute {
+    named!(doc_attr -> Attribute, call!(Attribute::parse_outer));
+    doc_attr.parse_str(&quote!(#[doc=#s]).to_string()).unwrap()
 }